feat: 기능에 따른 JS 파일 모듈화
